+++ neo/idlib/math/Matrix.h
@@ -2280,7 +2280,7 @@ ID_INLINE void idMatX::SetData( int rows, int columns, float *data ) {
 	if ( mat != NULL && alloced != -1 ) {
 		Mem_Free16( mat );
 	}
-	assert( ( ( (int) data ) & 15 ) == 0 ); // data must be 16 byte aligned
+	assert( ( ( (uintptr_t) data ) & 15 ) == 0 ); // data must be 16 byte aligned
 	mat = data;
 	alloced = -1;
 	numRows = rows;
