+++ neo/sys/linux/sound_alsa.cpp
@@ -305,7 +305,7 @@ void idAudioHardwareALSA::Write( bool flushing ) {
 		return;
 	}
 	// write the max frames you can in one shot - we need to write it all out in Flush() calls before the next Write() happens
-	int pos = (int)m_buffer + ( MIXBUFFER_SAMPLES - m_remainingFrames ) * m_channels * 2;
+	uintptr_t pos = (uintptr_t)m_buffer + ( MIXBUFFER_SAMPLES - m_remainingFrames ) * m_channels * 2;
 	snd_pcm_sframes_t frames = id_snd_pcm_writei( m_pcm_handle, (void*)pos, m_remainingFrames );
 	if ( frames < 0 ) {
 		if ( frames != -EAGAIN ) {
